only do renaming
