+++ src/lightly_train/_task_models/dinov2_linear_semantic_segmentation/train_model.py
@@ -30,7 +30,7 @@ from lightly_train._metrics.semantic_segmentation.task_metric import (
 )
 from lightly_train._optim import optimizer_helpers
 from lightly_train._task_models.dinov2_linear_semantic_segmentation.task_model import (
-    DINOv2LinearSemanticSegmentation,
+    LinearSemanticSegmentation as DINOv2LinearSemanticSegmentation,
 )
 from lightly_train._task_models.dinov2_linear_semantic_segmentation.transforms import (
     DINOv2LinearSemanticSegmentationTrainTransform,
